test(topup): implement test cases for the topup feature
