+++ circuits/ts/__tests__/ProcessMessages.test.ts
@@ -427,4 +427,74 @@ describe("ProcessMessage circuit", function test() {
       });
     });
   });
+
+  describe("1 user, 1 topup, 2 messages", () => {
+    const maciState = new MaciState(STATE_TREE_DEPTH);
+    const voteOptionIndex = BigInt(0);
+    let stateIndex: bigint;
+    let pollId: number;
+    let poll: Poll;
+    const userKeypair = new Keypair();
+
+    before(() => {
+      // Sign up and publish
+      stateIndex = BigInt(
+        maciState.signUp(userKeypair.pubKey, voiceCreditBalance, BigInt(Math.floor(Date.now() / 1000))),
+      );
+
+      pollId = maciState.deployPoll(
+        BigInt(Math.floor(Date.now() / 1000) + duration),
+        maxValues,
+        treeDepths,
+        messageBatchSize,
+        coordinatorKeypair,
+      );
+
+      poll = maciState.polls[pollId];
+    });
+
+    it("should work when publishing 2 vote messages and a topup (the second vote uses more than initial voice credit balance)", async () => {
+      // First command (valid)
+      const command1 = new PCommand(
+        stateIndex, // BigInt(1),
+        userKeypair.pubKey,
+        voteOptionIndex + 1n, // voteOptionIndex,
+        5n, // vote weight
+        BigInt(2), // nonce
+        BigInt(pollId),
+      );
+
+      const signature1 = command1.sign(userKeypair.privKey);
+
+      const ecdhKeypair1 = new Keypair();
+      const sharedKey1 = Keypair.genEcdhSharedKey(ecdhKeypair1.privKey, coordinatorKeypair.pubKey);
+      const message1 = command1.encrypt(signature1, sharedKey1);
+
+      poll.publishMessage(message1, ecdhKeypair1.pubKey);
+
+      poll.topupMessage(new Message(2n, [BigInt(stateIndex), 50n, 0n, 0n, 0n, 0n, 0n, 0n, 0n, 0n]));
+
+      // First command (valid)
+      const command = new PCommand(
+        stateIndex, // BigInt(1),
+        userKeypair.pubKey,
+        voteOptionIndex, // voteOptionIndex,
+        10n, // vote weight
+        BigInt(1), // nonce
+        BigInt(pollId),
+      );
+
+      const signature = command.sign(userKeypair.privKey);
+
+      const ecdhKeypair = new Keypair();
+      const sharedKey = Keypair.genEcdhSharedKey(ecdhKeypair.privKey, coordinatorKeypair.pubKey);
+      const message = command.encrypt(signature, sharedKey);
+
+      poll.publishMessage(message, ecdhKeypair.pubKey);
+
+      const inputs = poll.processMessages(pollId);
+      const witness = await circuit.calculateWitness(inputs, true);
+      await circuit.checkConstraints(witness);
+    });
+  });
 });
