+++ core/package.json
@@ -7,8 +7,8 @@
   "scripts": {
     "watch": "tsc --watch",
     "build": "tsc",
-    "test-processMessage": "ts-mocha --exit ts/__tests__/ProcessMessage.test.ts",
-    "test-maciState": "ts-mocha --exit ts/__tests__/MaciState.test.ts",
+    "test:e2e": "ts-mocha --exit ts/__tests__/e2e.test.ts",
+    "test:unit": "ts-mocha --exit ts/__tests__/MaciState.test.ts",
     "test": "nyc ts-mocha --exit ts/__tests__/*.test.ts"
   },
   "dependencies": {
